Move all of the document block definitions into a separate file
This greatly improves the readability of the code since it helps
tidy up the main document class where the real work is done. The
namespaces are unchanged.
